[320]: Fix posts server rendering
